# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "CONTACTS REST API",
    "description": ""
  },
  "host": "",
  "basePath": "",
  "tags": [
    {
      "name": "",
      "description": ""
    }
  ],
  "schemes": [],
  "securityDefinitions": {},
  "paths": {
    "/": {
      "get": {
        "summary": "heartbeat check",
        "description": "heartbeat check.",
        "produces": [
          "text/css"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "summary": "returns all the contacts in the mongoDB",
        "description": "returns all the contacts in the mongoDB",
        "parameters": [],
        "responses": {
          "200": {
            "description": "contacts json"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}